Split parsed logfile jsons, added radarid to script logs
Modified parse_dtaflow_log.py to write two separate json files: one for site scripts, and one for campus scripts. This is done with the goal of making it easier to separate the two when uploading them to the dashboard.
Also added RADARID to the summary logs for each script so it can easily be parsed if needed in the future.